Added directory support!
I still need to clean up exception handling and
provide instructive error messages when it occurs.

I should also clean up some of the code in places. It
isn't very pretty wrap a Runnable anonymous class in a
try statement, which we then add to the executor service.

Either each should get its own wrapper method to make it
look pretty, and or I should make a separate classes
for the anonymous ones, or just wrap them in a function themselves
(say, the function just returns the Runnable anonymous class).